Add CSP via <meta> tag
A restrictive CSP is possible since we don't rely on many external
parties at all. Fonts will be moved to another host in #30 so then
we can also remove that origin from the CSP.

Debated a little whether to allow `unsafe-inline` for scripts to
keep simple JavaScript, but that doesn't seem worth it. I was on
the fence about creating an external .js file over a script block
and a hash in the CSP, but in the end felt an extra network request
for our visitors is not worth it. For now.

Fixes #37